feat(challenges): Change [challengeId] page to SSR and move other CSR logics to a new component
